streamlit app/requirements.txt
streamlit==0.71.0
pandas==1.0.5
plotly==4.14.3
psycopg2-binary==2.9.3
